Refactor styles and better support for automatic dark mode
