--- src/assets/style.css
@@ -1,186 +1,0 @@
-:root {
-	--bs-body-bg: #ebebeb;
-}
-
-.navbar {
-	padding: 20px 40px 20px 40px;
-}
-
-.search-field {
-	position: relative;
-}
-
-.search-icon {
-	position: absolute;
-	padding: 10px;
-	pointer-events: none;
-	left: 0px;
-}
-
-.search-input {
-	padding-left: 30px;
-}
-
-.search-results {
-	position: absolute;
-	background-color: white;
-	width: 100%;
-
-	box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.2);
-	border-bottom-left-radius: 2px;
-	border-bottom-right-radius: 2px;
-
-	/*
-	 Opening/closing animations (having a delay before the results are hidden may
-	 fix a link-activation issue on some platforms).
-	 */
-	opacity: 1;
-	transform: scale(1, 1);
-	transform-origin: top;
-	transition: 0.25s ease opacity, 0.25s ease transform;
-}
-
-.search-field:not(:focus-within) .search-results {
-	transform: scale(1, 0);
-	opacity: 0;
-}
-
-.search-results ul {
-	list-style: none;
-	padding-left: 0;
-}
-
-.search-results ul :nth-child(2n) {
-	background-color: rgba(100, 100, 100, 0.01);
-}
-
-.search-results li {
-	padding-left: 20px;
-	padding-right: 20px;
-	padding-top: 2px;
-	padding-bottom: 5px;
-}
-
-.search-results li .icon {
-	padding-right: 5px;
-}
-
-.search-results li a {
-	display: flex;
-	flex-direction: row;
-	text-decoration: none;
-}
-
-.search-results li a:hover {
-	text-decoration: underline;
-}
-
-.banner {
-	padding-top: 4rem;
-	padding-bottom: 4rem;
-}
-
-.nav-pills {
-	--bs-nav-pills-border-radius: 0.375rem;
-	--bs-nav-pills-link-active-color: var(--bs-black);
-	--bs-nav-pills-link-active-bg: var(--bs-gray-300);
-}
-
-.nav {
-	--bs-nav-link-color: var(--bs-gray-700);
-	--bs-nav-link-hover-color: var(--bs-black);
-	--bs-nav-link-disabled-color: #6c757d;
-}
-
-.nav-list {
-	background-color: white;
-	border-radius: 0.375rem;
-}
-
-.tab-content {
-	min-height: auto;
-	background-color: white;
-	border-radius: 0.375rem;
-}
-
-.anchor {
-	position: relative;
-	top: -100px;
-	display: block;
-	height: 0;
-	overflow: hidden;
-}
-
-.link-card {
-	list-style: none;
-	display: flex;
-	border: 1px solid hsla(240, 6%, 83%, 0.63);
-	border-radius: 0.5rem;
-	min-height: 200px;
-	position: relative;
-	line-height: 1.4;
-	padding: 1em 1.3em;
-	border-radius: 0.35rem;
-	color: var(--text-color);
-	background-image: linear-gradient(120deg, #fff 45%, #f5f5f5);
-	display: block;
-}
-
-.link-card:is(:hover, :focus-within) {
-	filter: drop-shadow(0 10px 10px rgba(76, 76, 109, 0.07059));
-}
-
-.plugin-description {
-	display: -webkit-box;
-	-webkit-box-orient: vertical;
-	-webkit-line-clamp: 2;
-	overflow: hidden;
-}
-
-.plugin-title {
-	text-decoration: none;
-	color: black;
-}
-
-.plugin-category {
-	text-transform: capitalize;
-	margin-top: -4px;
-}
-
-.plugin-recommended {
-	text-transform: capitalize;
-	margin-top: -4px;
-	background-color: rgb(255, 217, 47);
-}
-
-.plugin-action {
-	position: absolute;
-	bottom: 1em;
-	left: 1.425em;
-	right: 0;
-	margin: 0;
-}
-
-.wrap {
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
-.horizontal-scroll-wrapper {
-	white-space: nowrap;
-	overflow-x: scroll;
-}
-
-.screenshot-placeholder {
-	background-color: #ebebeb;
-	border-radius: 0.5rem;
-	display: inline-block;
-	margin: 0.5em 0.5em 0.5em 0;
-}
-
-#screenshots-container {
-	#raw-screenshots > img {
-		max-width: 100%;
-	}
-}
