+++ src/components/head-config.mustache
@@ -9,4 +9,3 @@
 	window.siteRoot = `{{&config.site}}`;
 </script>
 <script src="{{&config.site}}/bundle.js"></script>
-<link rel="stylesheet" href="{{&config.site}}/style.css"/>
