+++ src/components/navabar.mustache
@@ -1,4 +1,4 @@
-<nav class="navbar navbar-expand-lg bg-light sticky-top">
+<nav class="navbar navbar-expand-lg bg-body-tertiary sticky-top">
 	<div class="container">
 		<a class="navbar-brand" href="{{&config.site}}">
 			<img src="{{&config.site}}/joplin-logo.png" width="110" alt="Home"/>
