+++ src/runtime/search/initializeSearch.ts
@@ -1,4 +1,4 @@
-import type PluginDataManager from './PluginDataManager';
+import type PluginDataManager from '../PluginDataManager';
 
 const initializeSearch = (
 	pluginData: PluginDataManager,
