+++ src/runtime/search/style.scss
@@ -1,0 +1,68 @@
+.search-field {
+	position: relative;
+}
+
+.search-icon {
+	position: absolute;
+	padding: 10px;
+	pointer-events: none;
+	left: 0px;
+}
+
+.search-input {
+	padding-left: 30px;
+}
+
+.search-results {
+	position: absolute;
+	background-color: var(--bs-tertiary-bg);
+	width: 100%;
+
+	box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.2);
+	border-bottom-left-radius: 2px;
+	border-bottom-right-radius: 2px;
+
+	/*
+	 Opening/closing animations (having a delay before the results are hidden may
+	 fix a link-activation issue on some platforms).
+	 */
+	opacity: 1;
+	transform: scale(1, 1);
+	transform-origin: top;
+	transition: 0.25s ease opacity, 0.25s ease transform;
+}
+
+.search-field:not(:focus-within) .search-results {
+	transform: scale(1, 0);
+	opacity: 0;
+}
+
+.search-results ul {
+	list-style: none;
+	padding-left: 0;
+}
+
+.search-results ul :nth-child(2n) {
+	background-color: rgba(100, 100, 100, 0.01);
+}
+
+.search-results li {
+	padding-left: 20px;
+	padding-right: 20px;
+	padding-top: 2px;
+	padding-bottom: 5px;
+}
+
+.search-results li .icon {
+	padding-right: 5px;
+}
+
+.search-results li a {
+	display: flex;
+	flex-direction: row;
+	text-decoration: none;
+}
+
+.search-results li a:hover {
+	text-decoration: underline;
+}
